Avoid complex dependencies for TrackedMetrics
