+++ InnerEye/ML/utils/training_util.py
@@ -10,43 +10,8 @@ import numpy as np
 import torch
 
 from InnerEye.Common import common_util
-from InnerEye.Common.metrics_dict import MetricsDict
 from InnerEye.Common.type_annotations import DictStrFloat
-
-
-@dataclass
-class ModelForwardAndBackwardsOutputs:
-    """
-    Dataclass to store results from a single model forward and backwards pass.
-    """
-    loss: float
-    logits: Union[torch.Tensor, np.ndarray]
-    labels: Union[torch.Tensor, np.ndarray]
-
-
-@dataclass
-class ModelOutputsAndMetricsForEpoch:
-    """
-    Dataclass to store results from a single epoch.
-    """
-    metrics: MetricsDict
-    model_outputs: List[ModelForwardAndBackwardsOutputs]
-    is_train: bool
-
-    def __post_init__(self) -> None:
-        common_util.check_properties_are_not_none(self, ignore=["metrics"])
-
-    def get_logits(self) -> torch.Tensor:
-        """
-        Get concatenated logits from each training/validation batch in this epoch.
-        """
-        return torch.cat([x.logits for x in self.model_outputs])
-
-    def get_labels(self) -> torch.Tensor:
-        """
-        Get concatenated labels from each training/validation batch in this epoch.
-        """
-        return torch.cat([x.labels for x in self.model_outputs])
+from InnerEye.ML.metrics_dict import MetricsDict
 
 
 @dataclass(frozen=True)
