+++ InnerEye/ML/visualizers/grad_cam_hooks.py
@@ -10,7 +10,7 @@ import numpy as np
 import torch
 from torch.nn import Module
 
-from InnerEye.Common.metrics_dict import SEQUENCE_POSITION_HUE_NAME_PREFIX
+from InnerEye.Common.metrics_constants import SEQUENCE_POSITION_HUE_NAME_PREFIX
 from InnerEye.ML.dataset.scalar_sample import ScalarItem
 from InnerEye.ML.dataset.sequence_sample import ClassificationItemSequence
 from InnerEye.ML.models.architectures.classification.image_encoder_with_mlp import ImagingFeatureType
@@ -506,7 +506,7 @@ class VisualizationMaps:
         for label_index in range(len(target_indices)):
             target_position = target_indices[label_index]
             current_output_dir = self.config.visualization_folder / f"{SEQUENCE_POSITION_HUE_NAME_PREFIX}_" \
-                f"{target_position}"
+                                                                    f"{target_position}"
             current_output_dir.mkdir(exist_ok=True)
             guided_grad_cams, grad_cams, pseudo_cam_non_img, probas = self.generate(input_batch,
                                                                                     target_position,
