+++ Tests/Common/test_metrics_dict.py
@@ -12,14 +12,14 @@ import pandas as pd
 import pytest
 from sklearn.metrics import roc_auc_score, roc_curve
 
-from InnerEye.Common.metrics_dict import DataframeLogger, Hue, MetricType, MetricsDict, PredictionEntry, \
-    ScalarMetricsDict, SequenceMetricsDict, average_metric_values
+from InnerEye.Common.metrics_constants import LoggingColumns, MetricType
 from InnerEye.Common.output_directories import OutputFolderForTests
 from InnerEye.ML import metrics
 from InnerEye.ML.common import ModelExecutionMode
 from InnerEye.ML.config import BACKGROUND_CLASS_NAME
+from InnerEye.ML.metrics_dict import DataframeLogger, Hue, MetricsDict, PredictionEntry, \
+    ScalarMetricsDict, SequenceMetricsDict, average_metric_values
 from InnerEye.ML.utils.io_util import tabulate_dataframe
-from InnerEye.ML.utils.metrics_constants import LoggingColumns
 
 
 def test_average_metric_values() -> None:
