+++ Tests/ML/models/test_2d_scalar_model.py
@@ -9,7 +9,7 @@ from pathlib import Path
 import pytest
 
 from InnerEye.Common.common_util import logging_to_stdout
-from InnerEye.Common.metrics_dict import MetricType
+from InnerEye.Common.metrics_constants import MetricType
 from InnerEye.Common.output_directories import OutputFolderForTests
 from InnerEye.ML import model_testing, model_training
 from InnerEye.ML.common import ModelExecutionMode
