+++ Tests/ML/reports/test_segmentation_report.py
@@ -7,11 +7,11 @@ from pathlib import Path
 
 import pandas as pd
 
+from InnerEye.Common.metrics_constants import MetricsFileColumns
 from InnerEye.Common.output_directories import OutputFolderForTests
 from InnerEye.ML.reports.notebook_report import generate_segmentation_notebook
 from InnerEye.ML.reports.segmentation_report import describe_score, worst_patients_and_outliers
 from InnerEye.ML.utils.csv_util import COL_IS_OUTLIER
-from InnerEye.ML.utils.metrics_constants import MetricsFileColumns
 
 
 def test_generate_segmentation_report(test_output_dirs: OutputFolderForTests) -> None:
